requirements.txt
aiohttp==3.8.1
aiofiles==0.7.0
numpy==1.21.2
pandas==1.3.3
python-dotenv==0.19.0
requests==2.26.0
scikit-learn==0.24.2
tensorflow==2.6.0
tqdm==4.62.3
PyQt5==5.15.4
cryptography==3.4.8
matplotlib==3.4.3
